+++ src/main.ts
@@ -2,25 +2,20 @@ import { createApp } from "vue";
 import { createPinia } from "pinia";
 import 'vuetify/styles';
 import { createVuetify } from 'vuetify';
-import { initializeApp } from "firebase/app";
-import { get, getDatabase, limitToFirst, ref, set, query, orderByKey, child, update} from "firebase/database";
 import { aliases, mdi } from 'vuetify/iconsets/mdi'
 import App from "@/App.vue";
 import '@mdi/font/css/materialdesignicons.css'
 import router from "@/router";
 
-const firebaseConfig = {
-  apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
-  authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-  projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-  storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_ID,
-  appId: import.meta.env.VITE_FIREBASE_APP_ID,
-  databaseURL: import.meta.env.VITE_FIREBASE_DATABASE_URL,
-};
 
-const firebaseApp = initializeApp(firebaseConfig);
-const DB = getDatabase(firebaseApp);
+const app = createApp(App);
+
+app.use(createPinia());
+app.use(router);
+app.use(vuetify);
+
+app.mount("#app");
+
 
 const vuetify = createVuetify({
   icons: {
@@ -31,15 +26,6 @@ const vuetify = createVuetify({
     }
   },
 });
-const app = createApp(App);
-
-app.use(createPinia());
-app.use(router);
-app.use(vuetify);
-
-app.mount("#app");
-
-
 
 /* var mockUsers = [
   {
